backend/requirements.txt
# ── Web framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Data processing ───────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── HTTP client (AI providers) ────────────────────────────
httpx==0.27.0

# ── Firebase Firestore ────────────────────────────────────
firebase-admin==6.5.0

# ── Config ────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1

# ── Cache ─────────────────────────────────────────────────
cachetools==5.3.3

# ── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
